sd: Try to fix PresenterPane visibility check
See previous commit

    Change-Id: I933c63303d1f055b4c0eb826f6726df3fe2a7033
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Tue May 13 23:38:34 2025 +0200

        sd: Use concrete rtl::Reference<PresenterPaneBorderPainter>

which made clear that PresenterPaneBase::IsVisible
was effectively always returning false.

Instead, try to query the `mxBorderWindow` for the XWindow2
interface, and check whether it is visible.

This should presumably be correct, and with a breakpoint set
there, the checks succeed when starting an Impress presentation
with the qt6 VCL plugin on Linux, but this is otherwise untested.
(If this commit results in any issues because of the
bounding box now actually getting updated and thus
invalidations taking place, maybe that bounding box/invalidation
logic should be refined or dropped altogether as it was presumably
logically dead before this commit.)

Change-Id: I91a45944e65983068463364d08cbc3c68f3a6ad6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185289
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>