tdf#166319 sw interdependent redlines: fix DOCX export of delete under format
The bugdoc has <del>AA<format>BB</format>CC</del> in it, DOCX import
works fine, but exporting back to DOCX results in a document Word can't
open.

It seems that the problem is: this is a "format on top of a delete"
redline, so we kept writing <w:t> for the format redline content, while
Word considers "format on top of a delete" essentially a type of delete,
so it requires <w:delText> instead.

Fix the problem by extending DocxAttributeOutput::RunText() to consider
both "delete" and "delete, then something on top of it" as a delete
redline, that results in the correct markup.

Keep our own DOCX import unchanged to still accept both <w:t> and
<w:delText> for text inside a delete redline.

Change-Id: Id64fcad322bc58dc0ef3e453445f83248d6f3cff
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185301
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins