tdf#88752 DOCX import: Ignore text grid base height inside tables
When laying out documents containing a CJK grid, Writer fixes line
heights to a multiple of the grid line pitch. Writer also applies this
correction to text inside tables.

Unfortunately, this is not the expected behavior for DOC/DOCX files. In
order to display such files correctly, this change updates Writer to
skip this line height adjustment when the MS_WORD_COMP_GRID_METRICS
compatibility flag is set.

Change-Id: I40792191a7690cb53d2ce02725a0a8a68dd8e606
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185307
Reviewed-by: Jonathan Clark <jonathan@libreoffice.org>
Tested-by: Jenkins