sd presenter a11y: Stop tracking a11y state in controller
Stop tracking whether a11y is active in
PresenterController. So far, this was the case once
PresenterAccessible::getAccessibleContext was called
for the first time, at which point that method was calling
PresenterController::SetAccessibilityActiveState.

Instead, let PresenterController call methods on its
a11y object unconditionally.
The object is already created unconditionally in
PresenterController::InitializeMainPane, and the
corresponding PresenterAccessible methods still don't
do much as long as PresenterAccessible::getAccessibleContext
hasn't been called, because PresenterAccessible::mpAccessibleConsole
is only created in there.

What's new is that PresenterController::notifyConfigurationChange
now unconditionally calls

    UpdateCurrentSlide(0);

for the ConfigurationUpdateEndEventType case, but making sure the
slide references are up to date seems reasonable to me.

Change-Id: Ifbc78578caf68323699b54b883e42ffd9295fff9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185325
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>