tdf#165918: speed up mark names uniqueness check
MarkManager::makeMark call can be quite expensive, when it needs to
ensure that the passed name is unique. It needs to iterate over all
the existing marks in m_vAllMarks, checking their names. At import
time, this creates quadratic complexity, even when the passed names
are already all unique.

This change introduces a dedicated performance mode, when names are
not deduplicated in load time. When the mode is ended, an optimized
loop is performed, using a set of names for much faster uniqueness
check. Depending on the number of bookmarks, I saw the improvement
of the test documents loading times up to 40% (like 190 s -> 115 s;
of course, realistic documents would have more modest improvements).

Change-Id: I4d624ea6cd7268e7bcfdefecac6d3f1bb58edc28
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185340
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>