tdf#76029 sd console: Scroll to new slide as needed
When switching to a new slide in Impress, make sure that
this new slide is fully visible in the Presenter Console
"Slides" mode automatically, by scrolling to adjust the
visible area as needed.
(Make the slide's row the topmost one that's visible.)

This removes the need for the user to manually scroll
using the scroll bar after switching slides.

Change-Id: I7d2a245e6f4b705a8a2492ccb02bbef490adbe88
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185358
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>