sd presenter a11y: Set preview's parent
The `mpAccessibleConsole` is the XAccessibleContext
for the PresenterAccessible, i.e. the PresenterAccessible
object is the XAccessible object for its `mpAccessibleConsole`.

PresenterAccessible::UpdateAccessibilityHierarchy sets
the `mpAccessiblePreview` as a child of `mpAccessibleConsole`.
Therefore, the PresenterAccessible is the parent XAccessible
for the preview. Set it accordingly.

This fixes an inconsistency in the a11y hierarchy that
could be seen as follows with the qt6 VCL plugin:

* start an Impress presentation with Presenter Console activated
* start Accerciser
* navigate to the "Presenter Console" object in Accerciser's
  tree view of LO a11y.
* in Accerciser's IPython console:

    In [30]: acc.get_child_at_index(0).get_parent() == acc
    Out[30]: False

This means the Presenter Console's a11y child at index 0 (which
is the current slide preview) doesn't report it's actual
parent as its parent.

With this commit in place, this behaves as expected now:

    In [1]: acc.get_child_at_index(0).get_parent() == acc
    Out[1]: True

Change-Id: Id41da5139214437de44d6b1cf29ad8de1d9d13be
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185364
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins