tdf#151336 editeng: Fix bad cursor pos when moving to RTL paragraph
This change fixes a bug causing incorrect visual cursor positioning when
advancing the cursor from an LTR paragraph to an RTL paragraph that
starts with an LTR run.

When advancing the cursor past the end of a paragraph, the cursor should
be placed at the conceptual start of the following paragraph.
Previously, Edit Engine did this by placing the cursor immediately
before the character at index 0. The result looks correct in many
situations, but it fails in the case of bidi text: the conceptual start
of an RTL paragraph is the right edge of the text area, but when the
paragraph begins with LTR characters, the character at position 0 may
start arbitrarily far to the left of the expected cursor position.

Edit Engine already contains logic to correctly position the cursor for
the home/end keys. This change extends that logic to include cases where
the cursor is moved between paragraphs.

Change-Id: I0d23b87beacd53d7508b74445d25f9024dbaf89c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185367
Tested-by: Jenkins
Reviewed-by: Jonathan Clark <jonathan@libreoffice.org>