sd presenter a11y: Drop indirection to set win accessible
So far the logic including indirection via UNO interfaces
was like this:

* The PresenterAccessible ctor calls XPane2::setAccessible,
  passing itself as the accessible.
* The XPane2 implementations call vcl::Window::SetAccessible
  with the accessible.
* The XPane2 implementations (at least the one in
  FullScreenPane::setAccessible) calls XInitialization::initialize
  on the accessible (which is implemented in PresenterAccessible)
  to set the window's a11y parent as the a11y parent of
  the PresenterAccessible.
* PresenterAccessible::initialize takes that parent and remembers
  it and/or sets it on the underlying console a11y object.

Get rid of that indirection by directly retrieving the
pane's vcl::Window in the PresenterAccessible ctor and
moving the logic there.

Linking sdui against the toolkit module is newly needed because
of the newly introduced use of VCLUnoHelper::GetWindow.

Stop implementing the XInitialization iface in PresenterAccessible.

Replace the XFrame2::setAccessible implementations with dummy
ones that trigger an assert for now.
(They will be dropped from the interface altogether in
a separate commit.)

PresenterAccessible no longer needs an XPane2 any more,
which allows to drop some UNO querying between XPane and XPane
interfaces.

Change-Id: I055dd19bbec810126d67e7795ffdfab257beeed3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185365
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>