svx: lower the priority of StylesListUpdateTask to TaskPriority::DEFAULT_IDLE
The update of styles previews is probably less important thatn the
update of the document area, so in case this stays in the
TaskPriority::HIGHEST -> TaskPriority::REPAINT region, that's
potentially problematic.

In practice for example a LOK client may decide to register its anyInput
callback to stop idle jobs in case no high priority jobs are scheduled
on the main loop, and this would delay paintTile() calls due to this
idle.

Change-Id: I6a3135caccb8abe9a8117e0d8c7bdda0df7e2780
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185369
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins