sd presenter a11y: Merge XAccessible{,Context} implementations
So far, PresenterAccessible was implementing the
XAccessible interface for the accessible object
used by PresenterController, and it was creating
another AccessibleObject (`mpAccessibleConsole`)
that implemented the XAccessibleContext interface
and was created in the call to
PresenterController:getAccessibleContext.

Stop that separation and let PresenterAccessible
also implement the XAccessibleContext interface,
and subclass AccessibleObject (which in turn
subclasses OAccessibleComponentHelper).

This aligns this further with what almost all
XAccessible/XAccessibleContext implementations
in the LO codebase do by now.

Apart from that, keep the logic mostly unchanged,
though the approach taken might be worth reconsidering
at some point in time in the future.
Currently, the way that announcement of the current
slide by screen readers (explicitly analyzed with Orca
on Linux with the gtk3 VCL plugin) works is by sending
a dummy focus event for the accessible object for the slide
preview of the current slide (see
`PresenterAccessible::NotifyCurrentSlideChangepreview`)
and then sending a name-changed event which is announced
by Orca because the "focused object"'s accessible name
changed. In reality, keyboard focus isn't actually on that preview
pane, however.
(A fair part of the UI also isn't accessible/not seen in Accerciser's
treeview of the LO a11y hierarchy, likely due to the fact that
custom button etc. implementations are used (see PresenterToolBar and
PresenterToolBar::Element and subclasses) instead of existing
vcl/weld widgets and their existing a11y implementations.)

Behavior with both, Orca and Linux with the gtk3 VCL plugin
and NVDA on Windows when switching slides in Presenter console,
as well as what was seen in Accerciser on Linux with gtk3 or qt6
was the same with this commit as prior to it in a quick test.

Change-Id: I0a347aba18d1a4e9c499c3438bcbe2143ca78f6f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185372
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins