NDOO: NotesDisplayOnOverlay
For Calc, showing Notes/Postit info on Mouse Hoovering was
done by painting directly to the SplitWindows and invalidating
there again. This created unneccessary repaints of the whole
sheet. It was even painted 'after' the existing Overlay was
painted to make that work/not collide.

This is a perfect task for the Overlay where this info can
be shown and hidden again without the need to repaint the
calc sheet at that places at all.

The new visualization does exactly that. It still is Timer-
Based to not trigger this all the time (timings unchanged),
but now creates the needed visualization on the Overlay.

While the visualization data was not changed, the mechanism
had to be changed to use transformations/overlay/primitives.

This was partially complicated due to the 'combined split
view' used to show that: The SdrCaptionObj can span over
multiple windows in split view, separated by the window
framing lines - it looks like a glass window with segments,
with the visualisation being shown 'behind' the cross.
Technically this means that the same visualisation data
has to be shown - dependent on the view split used - in
up to four windows. This just uses the same primitives
for visualisation embedded in the needed transformation
primitives.

Also the previous version created and destroyed a SdrModel
and SdrPage (complete with ItemPoos and ItemSets) for
every visualization, this is no longer needed, too.

Change-Id: I18c6d63d346ea89fca21a4d856ad66c82156a774
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185417
Reviewed-by: Armin Le Grand <Armin.Le.Grand@me.com>
Tested-by: Jenkins