tdf#166319 sw interdependent redlines: handle reject of delete under format
The bugdoc has <del>AA<format>BB</format>CC</del> in it, and once the
cursor is in the middle of the AA, then pressing Reject should get rid
of the delete and keep the format, but it resulted in no redlines and a
"AACC" string.

What happened is that we found the delete under format, but the action
was still reject, and it's OK to implement reject by deleting the
content for insert-then-format, but this won't work for
delete-then-format.

Fix the problem by reusing lcl_AcceptInnerInsertRedline() for this
purpose: accepting an underlying insert is effectively a NOP (body text
doesn't change), and rejecting a delete is in the same category.

Rename to lcl_DeleteInnerRedline(), because what this function does is
to get rid of "underlying" (earlier) redlines of a redline, so this is
useful for rejecting deletes as well, not just for accepting inserts.

Change-Id: Ie2db419844b77638d12fa6edfcbc0d5c3b59ad20
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185384
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>