LOCRDT sw: deactivate active comment when fly is selected
CppunitTest_sw_uiwriter5 CPPUNIT_TEST_NAME="testImageComment" fails
because it inserts a comment, putting the cursor into the comment, then
selecting an image, putting the sw cursor onto the SwGrfNode, but the
active comment remains active, hence later the selection type is PostIt,
not anything fly related as it should be.

This appears to be a general problem and the test worked by accident before.

Change-Id: Ic49b46d16cd2412e379015f10fe4fc7224efbd80
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185409
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>