tdf#166510 writerfilter: move empty sectPr bottom margin to prev para
This is follow-up to my 25.2.3 commit for tdf#165047:
    sw mso-compat layout: always consolidate top margin

The problem here was not the consolidation itself,
but the fact that the paragraph after the section page break
needs to consolidate based on the below spacing of the sectPr paragraph
and we simply deleted that info along with the empty sectPr.
The result was that consolidation was working with the wrong number.

Since the bottom margin of the previous paragraph
has no impact on the empty sectPr,
we can simply transfer the sectPr's spacing to the previous paragraph,
which of course now becomes the last paragraph before the page break.

Slightly unexpected were some examples where no spacing was identified
for the sectPr context. The unit test includes this complication...

make CppunitTest_sw_ooxmlexport25 \
    CPPUNIT_TEST_NAME=testTdf166510_sectPr_bottomSpacing

Change-Id: I9bcae239480035b02653ec93a37f6bf05d135d8f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185424
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins
(cherry picked from commit 5ff1fd13f1879105ecc0541c55f30cea32ae54f4)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185430
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>