tdf#130857 qt weld: Fall back to icons from system (theme)
If no icon for a given icon name can be found in the LibreOffice
icon themes, fall back to using non-LibreOffice icon themes, e.g.
the Breeze one used by default on KDE Plasma.

This is e.g. relevant for the "go-previous" icon used in Writer's
"Insert" -> "Footnote and Endnote" -> "Insert special Footnote/Endnote"
dialog.

Without this commit in place, this would trigger the

    assert(!aPixmap.isNull() && "Failed to create icon pixmap");

in the `QPixmap toQPixmap(const BitmapEx& rBitmapEx)` helper
in a WIP branch where support for that dialog is declared
for SAL_VCL_QT_USE_WELDED_WIDGETS=1.

The button using that icon isn't actually visible when
triggering the dialog as described above, but is when
right-clicking on an existing footnote anchor and selecting
the "Footnote/Endnote" context menu entry to open the dialog
for editing the existing footnote.

Change-Id: I47f6bae6978e05787d3de83ccce8660b6a5d8301
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185435
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>