tdf#130857 sw: Don't use "pack-type" of "end" in insertfootnote.ui
Avoid using that because it's no longer supported in GTK 4
and no special handling for it is implemented in the QtBuilder
either, see also

    commit 58544c8fec07b151bb74094e1a0be8b965caf1e9
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Sat Jan 18 18:05:39 2025 +0100

        tdf#130857 templatedlg.ui: Don't use "pack-type end"

for more background.

Explicitly set positions for elements in the dialog
to keep the visual order as it was.

No change in behavior intended or observed when testing
that dialog with the gtk3 or gen VCL plugin on Linux
for both scenarios mentioned in previous commit

    Change-Id: I47f6bae6978e05787d3de83ccce8660b6a5d8301
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Fri May 16 21:27:45 2025 +0200

        tdf#130857 qt weld: Fall back to icons from system (theme)

Change-Id: Ib57409e4360f6f2d4b0115d7ac8d5785cf010602
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185438
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>