tdf#130857 sw: Don't use GtkBox inside button box
At least the Qt equivalent of a GtkButtonBox, a
QDialogButtonBox, doesn't support "random" widgets,
but only subclasses of QAbstractButtonBox.
Therefore, place the "previous" and "next" buttons
directly into the button box without an extra
GtkBox parent.
This makes the visual appearance of those buttons
match the others in the dialog more, and makes them
show up in a WIP branch where support for that dialog
is declared when using native Qt widgets with
SAL_VCL_QT_USE_WELDED_WIDGETS=1 in the second
scenario of that dialog where the buttons are
displayed as described in previous commit

    Change-Id: I47f6bae6978e05787d3de83ccce8660b6a5d8301
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Fri May 16 21:27:45 2025 +0200

        tdf#130857 qt weld: Fall back to icons from system (theme)

Change-Id: I5f178f100727bb4f3c48127eace7e0b28d71229e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185439
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins