tdf#130857 qt weld: Report released mouse button in event
As the QSinglePointEvent::buttons doc [1] says:

> For mouse press, double click, or TabletPress events, this includes the
> button that caused the event.
>
> For mouse release or TabletRelease events, this excludes the button that
> caused the event.

However, LO expects the released mouse button to be set
in the MouseEvent passed to weld::Widget::signal_mouse_release
(in QtInstanceWidget::eventFilter).

Therefore, also include QSinglePointEvent::button (without an "s")
at the end, which "returns the button that caused the event." [2]

Without this commit in place, clicking into the table of available
characters in the special characters dialog (in a WIP branch
experimenting with adding support for that dialog for
SAL_VCL_QT_USE_WELDED_WIDGETS=1) was resulting in the
mouse getting grabbed, but the mouse kept being grabbed when the
left mouse button was released, meaning that e.g. clicking on buttons
in the dialog had no effect.

This is because the `rMEvt.IsLeft()` check in
SvxShowCharSet::MouseButtonUp returned false:

    bool SvxShowCharSet::MouseButtonUp(const MouseEvent& rMEvt)
    {
        if ( bDrag && rMEvt.IsLeft() )
        {
            // released mouse over character map
            if ( tools::Rectangle(Point(), GetOutputSizePixel()).Contains(rMEvt.GetPosPixel()))
                aSelectHdl.Call( this );
            ReleaseMouse();
            bDrag = false;
        }

        return true;
    }

[1] https://doc.qt.io/qt-6/qsinglepointevent.html#buttons
[2] https://doc.qt.io/qt-6/qsinglepointevent.html#button

Change-Id: I88f8314d037a1bcc462a5c4a6cdabd1f70d81bfe
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185440
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>