vcl: Move RoadmapWizard method impls to corresponding source file
Move these implementations of the vcl::Window based
RoadmapWizard to the existing source file where the
other methods of that class are implemented, instead
of having some of them in the source file for the weld::Widget
based WizardMachine, while others are elsewhere.

Move definition of two strings needed by both of them to
a header that both already include.

Change-Id: I8ece77cbc9d1d8a52b1e3dcedb7b70ac8a310eba
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185460
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>