vcl: Move RoadmapWizardMachine to own source file
Don't intersperse the method implementations of the weld::Widget
based RoadmapWizardMachine with those for the vcl::Window based
RoadmapWizard, but move the RoadmapWizardMachine ones to a separate
source file, to make the distinction clearer.

Also rename the roadmapwizard.hxx header to roadmapwizardmachine.hxx
accordingly, as it contains the declaration of the RoadmapWizardMachine
class whose method implementatons are now in the new source file
called roadmapwizardmachine.cxx.

Change-Id: I4aa2933c496c0a26a70ef3027f45c87d1a0b61bc
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185466
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>