tdf#166620: don't strip leading tab from footnotes/endnotes anymore
Before the fix for tdf#159382, the layout of footnote/endnote number
was incompatible between Writer and Word; in Writer, a spacing was
always created between them, while in Word, there was none. This
meant, that to create a Word file format with a footnote/endnote
that looked similar to what Writer showed, we needed to put a tab
in the beginning of the exported note text; and on import, we needed
to strip it away.

The fix for tdf#159382 introduced NoGapAfterNoteNumber compatibility
flag, which avoids the extra space between the note's number and its
text. This flag it active for all Word file formats; so we need to
keep the tabs in the beginning of the note text, or else we show it
wrong.

This change drops the code that stripped the tabs, reverting part of
commit fcbc2fc4ae4ea9a499616bfeaa6a2cbece93f01a (INTEGRATION: CWS
tuamfilterteam23 (1.95.10); FILE MERGED, 2004-02-26), and also commits
b38629ae210b204a6d24d6e9c5c62eaaf563d494 (cp#1000017 DOCX/RTF import:
avoid fake tab char in footnotes, 2013-12-05),
946fee3ef1e319ad63a599b72dbd55ef52cbc640 (tdf#106062 ooxmlimport: skip
fake tab only on hanging indent, 2018-04-24), and
a3783c0af4bd21eb9c001aadc60c660c06a47779 (tdf#106062 ww8import: skip
fake tab only on hanging indent, 2018-04-28).

Also, it makes sure to avoid outputting a tab from a document with the
compatibility flag set, even when the note doesn't start with a tab.

Change-Id: I8fa085d3ba45367476853723c6ccba1b047eb092
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185471
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Signed-off-by: Xisco Fauli <xiscofauli@libreoffice.org>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185499