tdf#166319 sw interdependent redlines: handle ODF import of insert under format
Open the bugdoc, the ODT file has an insert with a format on top of the
middle, but the format is lost in Writer.

It seems what happens is that the XML parser correctly populates the
RedlineInfo structure, but then XMLRedlineImportHelper::ConvertRedline()
decides to throw away the underlying insert using an explicit condition
that was added in commit 52d244dee88b111631680d8cd4c8b922f9640c15 (-
added: redline import, 2001-01-10).

Fix this by extending it similar to the accept redline's
CanCombineTypesForAcceptReject() by adding a CanCombineTypesForImport()
that now accepts insert-then-format, too.

This is similar to what was fixed for DOCX in commit
ed8257a67d26083c2c6cd60ecac331c9e0df15ed (tdf#166319 sw interdependent
redlines: handle format on top of insert, 2025-04-29). The ODT export
was working already as-is.

Change-Id: I2163b0350ca56264ebed5abd54735ef4d91ee2ad
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185500
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins