vcl: Have only one pointer to ORoadmap
So far, RoadmapWizard had two pointers to the ORoadmap
it uses (both pointing to the same object):

RoadmapWizard::mpViewWindow had a vcl::Window pointer to
the same object as the RoadmapWizardImpl::pRoadmap
of its `m_xRoadmapImpl` was pointing to (both set in
the ctor and disposed/unset in RoadmapWizard::dispose).

Rename RoadmapWizard::mpViewWindow to
RoadmapWizard::mpRoadmap and let it be the one and
only pointer to the OWizard.

Change-Id: I8d2becd4c9d84883d00222cccd35a9402df38ca5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185524
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>