vcl: Don't use RoadmapTypes in RoadmapWizardMachine
RoadmapTypes::ItemIndex is a sal_Int32 typedef used
primarily for the vcl (i.e. vcl::Window based) ORoadmap.

Don't use that in the weld::Widget based
RoadmapWizardMachine for no apparent reason, but use
other integer types appropriate for the context instead,
which makes clearer this class is not directly related to
ORoadmap (and doesn't make any direct use of vcl::Window
anywhere, but makes use of the weld abstraction).

Change-Id: I8a94e9d24abe6c4ec637f459094b9f8dbcd42992
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185526
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins