tdf#166652: do not use FormatEmpty for proportional spacing < 100%
The FormatEmpty optimization can't handle this spacing mode, it uses
the text height without modification.

The change in testTdf143384_tableInFoot_negativeMargins (fixing the
test checking incorrect page number) is directly related: there are
empty paragraphs there, with spacing formatted to 75% - ignoring
which was the reason of excessive height causing an extra page.

Change-Id: Ibfae8b102f88656a1b703cc73b71ab6ef28a3ef4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185542
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>