tdf#166319 sw interdependent redlines: handle ODF export of delete under format
The bugdoc has <del>AA<format>BB</format>CC</del> in it, DOCX import
works, but exporting to ODT turns the format-on-delete redline into a
plain "format" redline.

What happens is that various code only checks the m_eType of the first
SwRedlineData of an SwRangeRedline, so format-on-delete is handled as
format. But this is not correct: if the format-on-delete is to be
accepted, then it's just a small detail that there was a format on top
of it: the result should be text deleted from the document on accept,
not accepting the the format changes and leaving the text in the
document.

Fix the problem by:
1) Extending SwRangeRedline::Hide() and SwRangeRedline::Show(), so when
   the ODF export moves delete redlines to their own toplevel SwNode
   section, so it can refer to these ranges using a single SwStartNode
   index, this move happens for format-on-delete as well, not just deletes.
2) Then can extend SwXRedlinePortion::GetPropertyValue()'s
   lcl_GetSuccessorProperties() to produce an XText for the deleted
   range of a format-on-delete redline.
3) Finally consume this new XText in
   XMLRedlineExport::ExportChangeInfo().

Note that this way a format redline can now have its own XText (in case
it's on top of a delete), but we want to export the text only once, so
restrict the content export for a redline to the 'delete' type.

Change-Id: I91902a2ae4b464ace41717a196d19530be9d84d1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185548
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>