tdf#148703 Fix Writer crash on undo after paste
When copy/pasting a selection from a source document with a
custom page style and a header/footer over a selection in a
destination document that does not have that custom page
style, the undo information was not recording the insertion
of the header/footer nodes into the destination document.

Then, when undoing the copy/paste, the header/footer nodes
remained in the destination document and messed up the
absolute indexing of the insertion point of the overwritten
destination selection, so re-inserting the destination
selection during the undo causes a crash.

This change modifies SwDoc::CopyPageDescHeaderFooterImpl to
record the insertion of the header/footer nodes so that they
are correctly undone and don't mess up the indexing.

This requires a new undo class SwUndoCopyHeaderFooter.

There is still a bit of work needed after this change; the
custom page style will be recreated correctly, but it still
will not be applied to the pages that should have it in the
destination document after the copy is undone and then
redone. See the ticket for more details.

Test plan:

I added a regression test based on a pair of minimal docs:

CPPUNIT_TEST_NAME="SwCoreUndoTest testTdf148703CopyFooterAcrossDocuments" make CppunitTest_sw_core_undo

I have also manually tested against the example docs from
this comment:

https://bugs.documentfoundation.org/show_bug.cgi?id=148703#c0

Change-Id: I8a07cd0bff45c50487d61953b1f66cd7c05905dd
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185940
Code-Style: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>