tdf#166319 sw interdependent redlines: handle ODF import of delete under format
The bugdoc has <del>AA<format>BB</format>CC</del> in it, ODT import
created a model that contains <del>AACC</del> instead, i.e. the
format-on-delete part was lost.

The first problem was that XMLRedlineImportHelper::ConvertRedline()
rejected the format-on-delete, fixing the used
CanCombineTypesForImport() resulted in a format-on-delete redline that
was in the document, but the order was wrong: BBAACC (matching the order
in the file, since deletions are listed separately, before the body
text).

The second problem was that XMLRedlineImportHelper::InsertIntoDocument()
didn't consider format-on-delete when it looked for deletions: fixing
that results in the wanted AABBCC string in the document model.

With this, the ODF import/export for ins-then-del, ins-then-fmt &
del-then-fmt should work.

Change-Id: I7bcdd4d7452d9bca0ff31ce99cbc29795eb63fa5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185588
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>