sw a11y test: Don't rely on XWindow being XAccessible
The fact that the VCLXWindow (which implements the
XWindow interface) is currently also the XAccessible
for most vcl::Window subclasses (see Window::GetAccessible
and the default implementation in Window::CreateAccessible,
but which is overriden in some vcl::Window subclasses) is an
implementation detail and subject to change, so don't rely on that.
Instead, call Window::GetAccessible on the underlying
vcl::Window.

See also

    commit b608604d0bbecdc24a17175374581b4e2d720658
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu Dec 5 14:07:52 2024 +0000

        tdf#164093 tdf#157001 wina11y: Use vcl::Window's actual XAccessible

Change-Id: Id81ab5f90955ecc600e179164b5f9c7a771182d1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185597
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>